Develop

+++ src/ffte/ffte_cooleytukey.c
@@ -8,7 +8,7 @@
 
 void swap(double *X, uint64_t i, uint64_t j);
 
-void ffte_cooleytukey(double* x_real, double* x_imag, size_t N, unsigned char only_real_input, unsigned char inverse)
+void ffte_cooleytukey(double* x_real, double* x_imag, unsigned int N, unsigned char only_real_input, unsigned char inverse)
 {
 	int64_t M = log((double)N) / log(2.0);
